fix: refactor
- move operator map in separate file
- remove unnecessary code
- organize functions
